# chore: optimize segment producing