--- a/examples/bench/compare/segment/go.mod
+++ b/examples/bench/compare/segment/go.mod
@@ -2,7 +2,7 @@ module segment
 
 go 1.24.2
 
-require github.com/segmentio/kafka-go v0.4.47
+require github.com/segmentio/kafka-go v0.4.48
 
 require (
 	github.com/klauspost/compress v1.18.0 // indirect