--- a/examples/bench/compare/segment/main.go
+++ b/examples/bench/compare/segment/main.go
@@ -5,7 +5,9 @@ import (
 	"crypto/tls"
 	"flag"
 	"fmt"
+	"log/slog"
 	"os"
+	"slices"
 	"strconv"
 	"strings"
 	"sync"
@@ -21,12 +23,12 @@ var (
 	seedBrokers = flag.String("brokers", "localhost:9092", "comma delimited list of seed brokers")
 	topic       = flag.String("topic", "", "topic to produce to or consume from")
 
-	writeSize   = flag.Int("records-per-write", 1, "number of records to create before a single WriteMessage call")
-	batchSize   = flag.Int("batch-size", 1, "value for the kafka.Writer.BatchSize field")
-	acks        = flag.Int("acks", -1, "value for the Writer.Acks field")
-	recordBytes = flag.Int("record-bytes", 100, "bytes per record (producing)")
-	compression = flag.String("compression", "none", "compression algorithm to use (none,gzip,snappy,lz4,zstd, for producing)")
-	poolProduce = flag.Bool("pool", false, "if true, use a sync.Pool to reuse record slices (producing)")
+	batchSize           = flag.Int("batch-size", 10000, "value for the kafka.Writer.BatchSize field")
+	acks                = flag.Int("acks", 0, "value for the Writer.Acks field")
+	recordBytes         = flag.Int("record-bytes", 100, "bytes per record (producing)")
+	compression         = flag.String("compression", "none", "compression algorithm to use (none,gzip,snappy,lz4,zstd, for producing)")
+	poolProduce         = flag.Bool("pool", false, "if true, use a sync.Pool to reuse record slices (producing)")
+	noIdempotentMaxReqs = flag.Int("max-inflight-produce-per-broker", 5, "if idempotency is disabled, the number of produce requests to allow per broker")
 
 	consume = flag.Bool("consume", false, "if true, consume rather than produce")
 	group   = flag.String("group", "", "if non-empty, group to use for consuming rather than direct partition consuming (consuming)")
@@ -111,7 +113,7 @@ func main() {
 			Topic:        *topic,
 			BatchSize:    *batchSize,
 			RequiredAcks: kafka.RequiredAcks(*acks),
-			Async:        true,
+			Async:        false, // batching requires this to be false
 			Completion: func(ms []kafka.Message, err error) {
 				chk(err, "produce err: %v", err)
 				for _, m := range ms {
@@ -138,16 +140,13 @@ func main() {
 			die("unrecognized compression %s", *compression)
 		}
 
+		ctx := context.Background()
+		producer := NewProducer(w, *batchSize)
+		go producer.Run(ctx, *noIdempotentMaxReqs)
+
 		var num int64
-		var msgs []kafka.Message
 		for {
-			msgs = msgs[:0]
-			for i := 0; i < *writeSize; i++ {
-				msgs = append(msgs, kafka.Message{Value: newValue(num)})
-			}
-			w.WriteMessages(context.Background(), kafka.Message{
-				Value: newValue(num),
-			})
+			producer.Produce(&kafka.Message{Value: newValue(num)})
 			num++
 		}
 
@@ -184,20 +183,78 @@ var p = sync.Pool{
 }
 
 func newValue(num int64) []byte {
-	var buf [20]byte // max int64 takes 19 bytes, then we add a space
-	b := strconv.AppendInt(buf[:0], num, 10)
-	b = append(b, ' ')
-
 	var s []byte
 	if *poolProduce {
 		s = *(p.Get().(*[]byte))
 	} else {
 		s = make([]byte, *recordBytes)
 	}
+	formatValue(num, s)
 
-	var n int
-	for n != len(s) {
-		n += copy(s[n:], b)
-	}
 	return s
+}
+
+func formatValue(num int64, v []byte) {
+	var buf [20]byte // max int64 takes 19 bytes, then we add a space
+	b := strconv.AppendInt(buf[:0], num, 10)
+	b = append(b, ' ')
+
+	n := copy(v, b)
+	for n != len(v) {
+		n += copy(v[n:], b)
+	}
+}
+
+//------------------------------------------------------------------------------
+
+type Producer struct {
+	wr           *kafka.Writer
+	ch           chan *kafka.Message
+	maxBatchSize int
+}
+
+func NewProducer(wr *kafka.Writer, maxBatchSize int) *Producer {
+	return &Producer{
+		wr:           wr,
+		ch:           make(chan *kafka.Message, 10*maxBatchSize),
+		maxBatchSize: maxBatchSize,
+	}
+}
+
+func (p *Producer) Produce(msg *kafka.Message) {
+	p.ch <- msg
+}
+
+func (p *Producer) Run(ctx context.Context, maxThreads int) {
+	ch := make(chan []*kafka.Message, maxThreads)
+
+	for range maxThreads {
+		go func() {
+			msgs := make([]kafka.Message, p.maxBatchSize)
+
+			for batch := range ch {
+				for i, msg := range batch {
+					msgs[i] = *msg
+				}
+				p.produceBatch(ctx, msgs[:len(batch)])
+			}
+		}()
+	}
+
+	batch := make([]*kafka.Message, 0, p.maxBatchSize)
+	for msg := range p.ch {
+		batch = append(batch, msg)
+		if len(batch) == p.maxBatchSize {
+			ch <- slices.Clone(batch)
+			batch = batch[:0]
+		}
+	}
+}
+
+func (p *Producer) produceBatch(
+	ctx context.Context, batch []kafka.Message,
+) {
+	if err := p.wr.WriteMessages(ctx, batch...); err != nil {
+		slog.Error("WriteMessages failes", slog.Any("err", err))
+	}
 }