… examples/bench/compare/segment/go.sum