Before:

```
go run . -topic foo -compression=snappy
0.42 MiB/s; 4.42k records/s
0.43 MiB/s; 4.51k records/s
0.43 MiB/s; 4.47k records/s
0.44 MiB/s; 4.60k records/s
0.46 MiB/s; 4.81k records/s
0.40 MiB/s; 4.24k records/s
0.47 MiB/s; 4.88k records/s
0.44 MiB/s; 4.65k records/s
```

After:

```
go run . -topic foo -compression=snappy
151.22 MiB/s; 1585.65k records/s
169.11 MiB/s; 1773.20k records/s
168.29 MiB/s; 1764.67k records/s
173.17 MiB/s; 1815.83k records/s
171.54 MiB/s; 1798.78k records/s
169.92 MiB/s; 1781.72k records/s
161.79 MiB/s; 1696.47k records/s
158.54 MiB/s; 1662.38k records/s
166.67 MiB/s; 1747.62k records/s
160.80 MiB/s; 1686.13k records/s
167.65 MiB/s; 1757.96k records/s
```

For some reason this is still slower than https://github.com/twmb/franz-go/pull/1064 by about 20MiB/s...